requirements.txt
# Core dependencies
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0

# PDF processing
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==11.3.0

# Text processing
beautifulsoup4==4.12.2
lxml==4.9.3
email-validator==2.1.0
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1

# AWS Bedrock integration
boto3==1.35.94
botocore==1.35.94

# Optional (not used in current implementation)
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.108.0
# streamlit==1.50.0
